requirements.txt
# Requirements for GitHub Actions Auto Scraper
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
gunicorn==21.2.0
geopy==2.4.1
